- Fix a severe bug caused by merging #87. The patch does not check the return   values from get_config_entries_size(), while the return values can easily be   negative (-PAL_ERROR_INVAL). Any usage of this function should carefully   check the return values and also pass the size into get_config_entries() to   prevent buffer overflow.
- Fix a minor bug in handling IPC disconnection. The callback ipc_child_exit()
  never receives a "term_signal" parameter. The callback is called when the
  pipe to a child process unexpectedly terminates, which can only mean the child
  process has crashed. I believe the proper signal to send is SIGKILL.

- Remove a compilation warning caused by type mismatch of __malloc().

+++ LibOS/shim/src/shim_init.c
@@ -433,7 +433,7 @@ int read_environs (const char ** envp)
 
 struct config_store * root_config = NULL;
 
-static void * __malloc (int size)
+static void * __malloc (size_t size)
 {
     return malloc(size);
 }
